fix: missing val set change

+++ src/main/java/io/zeebe/monitor/rest/dto/GenericKafkaRecord.java
@@ -61,7 +61,7 @@ public class GenericKafkaRecord {
         return sourceRecordPosition;
     }
 
-    public void setSourceRecordPosition(int sourceRecordPosition) {
+    public void setSourceRecordPosition(long sourceRecordPosition) {
         this.sourceRecordPosition = sourceRecordPosition;
     }
